requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.25.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.116.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.28
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.3.3
openai==1.108.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
streamlit==1.49.1
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
yarl==1.20.1
zstandard==0.25.0
